educa/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-braces==1.15.0
django-debug-toolbar==4.2.0
django-embed-video==1.4.9
django-redisboard==8.4.0
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
pymemcache==4.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
